H.W

+++ 03 - Programming Using C++ - Level 1/Numbers Summation H.W/Numbers Summation H.W.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int Num1 = 20, Num2 = 30, Num3 = 10;
+    short Num1 = 20, Num2 = 30, Num3 = 10;
 
     cout << Num1 << " +\n" << Num2 << " +\n" << Num3;
     cout << "\n___________________________________\n";
